refactor: postDetail.js, PostReport.js 분리
