+++ sseuda-react/src/pages/post/css/PostDetail.module.css
@@ -50,14 +50,13 @@
     margin: 50px 0;
 }
 
-/* 전역 또는 해당 컴포넌트 CSS */
 .popupOverlay {
     position: fixed;
     top: 0;
     left: 0;
     width: 100%;
     height: 100%;
-    background-color: rgba(0, 0, 0, 0.5); /* 어두운 배경 */
+    background-color: rgba(0, 0, 0, 0.5);
     display: flex;
     justify-content: center;
     align-items: center;
@@ -81,5 +80,9 @@
 }
 
 .popupButtons button {
+    width: 60px;
+    height: 30px;
+    border-radius: 5px;
+    background-color: #F5C3A4;
     margin-left: 10px;
 }
